Push nowej aplikacji

+++ 02-backend/zarzadanie-magazynem/src/main/java/com/kc6379/zarzadaniemagazynem/controller/RolesController.java
@@ -32,7 +32,7 @@ public class RolesController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<RolesDto> getRoles(@PathVariable Integer id){
+    public ResponseEntity<RolesDto> getRoles(@PathVariable Long id){
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .body(rolesService.getRoles(id));
